feat: 현재 위치 기반 액티비티 추천 슬라이드 구현
